tests/wpt/web-platform-tests/css/css-sizing/parsing/max-height-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Intrinsic & Extrinsic Sizing Test: parsing max-height with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-sizing-3/#propdef-max-height">
<meta name="assert" content="max-height supports only the grammar 'none | <length-percentage> | min-content | max-content | fit-content(<length-percentage>)'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("max-height", "auto");
test_invalid_value("max-height", "min-content max-content");

test_invalid_value("max-height", "-10%");
test_invalid_value("max-height", "-0.5em");
</script>
</body>
</html>
